+++ packages/atomic/src/components/facets/atomic-segmented-facet-scrollable/atomic-segmented-facet-scrollable.tsx
@@ -5,6 +5,17 @@ import {Button} from '../../common/button';
 
 type ArrowDirection = 'right' | 'left';
 
+/**
+ * @internal
+ * The 'atomic-segmented-facet-scrollable' component wraps around one or several 'atomic-segmented-facet' to provide horizontal scrolling capabilities.
+ * @part scrollable-container - The wrapper for the entire component including the horizontal-scroll container and the arrow buttons.
+ * @part horizontal-scroll - The scrollable container for the segmented facets.
+ * @part left-arrow-box - The left arrow box containing both the left arrow button and the fade.
+ * @part right-arrow-box - The right arrow box containing both the right arrow button and the fade.
+ * @part arrow-button - The arrow button used to scroll left or right.
+ * @part fade - The white to transparent gradient.
+ */
+
 @Component({
   tag: 'atomic-segmented-facet-scrollable',
   styleUrl: 'atomic-segmented-facet-scrollable.pcss',
@@ -76,24 +87,26 @@ export class AtomicSegmentedFacetScrollable {
     const isLeft: boolean = direction === 'left';
     return [
       <Button
-        part="arrow"
+        part={`${direction}-arrow-button`}
         style="square-neutral"
-        class={`flex shrink-0 basis-8 justify-center items-center rounded absolute z-10 w-10 top-0 bottom-0 ${this.renderDirectionClass(
-          direction
-        )}`}
+        class={`flex shrink-0 basis-8 justify-center items-center rounded absolute z-[1] w-10 top-0 bottom-0 ${
+          isLeft ? 'left-0' : 'right-0'
+        }`}
         ariaHidden="true"
+        tabIndex="-1"
         onClick={() => this.slideHorizontally(direction)}
       >
         <atomic-icon
+          part={`${direction}-arrow-icon`}
           class="w-3.5"
           icon={isLeft ? ArrowLeftIcon : ArrowRightIcon}
         ></atomic-icon>
       </Button>,
       <div
-        part="fade"
-        class={`w-16 h-10 absolute top-0  z-[5] pointer-events-none from-background-80 ${this.renderFadeClass(
-          direction
-        )}`}
+        part={`${direction}-fade`}
+        class={`w-16 h-10 absolute top-0  z-0 pointer-events-none from-background-80 ${
+          isLeft ? 'bg-gradient-to-r left-0' : 'bg-gradient-to-l right-0'
+        }`}
       ></div>,
     ];
   }