+++ packages/atomic/src/components/search/facets/atomic-facet/atomic-facet.tsx
@@ -13,7 +13,6 @@ import {
   FacetConditionsManager,
 } from '@coveo/headless';
 import {
-  Bindings,
   BindStateToController,
   InitializableComponent,
   InitializeBindings,
@@ -38,7 +37,6 @@ import {
   getFieldValueCaption,
 } from '../../../../utils/field-utils';
 import {Schema, StringValue} from '@coveo/bueno';
-import {registerFacetToStore} from '../../../../utils/store';
 import {Hidden} from '../../../common/hidden';
 import {
   FocusTarget,
@@ -46,6 +44,7 @@ import {
 } from '../../../../utils/accessibility-utils';
 import {MapProp} from '../../../../utils/props-utils';
 import {FacetValuesGroup} from '../facet-values-group/facet-values-group';
+import {Bindings} from '../../atomic-search-interface/atomic-search-interface';
 
 /**
  * A facet is a list of values for a certain field occurring in the results, ordered using a configurable criteria (e.g., number of occurrences).
@@ -233,7 +232,7 @@ export class AtomicFacet implements InitializableComponent, BaseFacet<Facet> {
     this.facet = buildFacet(this.bindings.engine, {options});
     this.facetId = this.facet.state.facetId;
     this.inititalizeDependenciesManager();
-    registerFacetToStore(this.bindings.store, 'facets', {
+    this.bindings.store.registerFacet('facets', {
       label: this.label,
       facetId: this.facetId!,
       element: this.host,