+++ packages/atomic/CHANGELOG.md
@@ -3,6 +3,95 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.77.2](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.77.1...@coveo/atomic@1.77.2) (2022-06-28)
+
+
+### Bug Fixes
+
+* **atomic:** encode colon for i18n namespace feature ([#2125](https://github.com/coveo/ui-kit/issues/2125)) ([2ac53be](https://github.com/coveo/ui-kit/commit/2ac53bedffc329c045ebb4d714050f7136a0e22e))
+* make lockfiles part of the PR generated files ([#2167](https://github.com/coveo/ui-kit/issues/2167)) ([0b1d12a](https://github.com/coveo/ui-kit/commit/0b1d12a5f47879583ace719f0ebf03f0490499a5))
+
+
+
+
+
+## [1.77.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.77.0...@coveo/atomic@1.77.1) (2022-06-27)
+
+
+### Bug Fixes
+
+* **atomic:** improved aria label for rating facet values ([#2166](https://github.com/coveo/ui-kit/issues/2166)) ([6b2c386](https://github.com/coveo/ui-kit/commit/6b2c386183c597a95150775e231daae8f5595550))
+
+
+
+
+
+# [1.77.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.76.2...@coveo/atomic@1.77.0) (2022-06-27)
+
+
+### Bug Fixes
+
+* **atomic:** improved focus style of outline buttons ([#2153](https://github.com/coveo/ui-kit/issues/2153)) ([21d249c](https://github.com/coveo/ui-kit/commit/21d249c2d0866b8604eb4381e0766b44275d4489))
+* **atomic:** improved search box aria label ([#2164](https://github.com/coveo/ui-kit/issues/2164)) ([64a3d0f](https://github.com/coveo/ui-kit/commit/64a3d0fc6573aaf63516c27aa47b44ced156fe89))
+* **atomic:** missing zh key ([#2154](https://github.com/coveo/ui-kit/issues/2154)) ([81438b9](https://github.com/coveo/ui-kit/commit/81438b9c723252d1960e7695c4b3d4d41e8010ca))
+* **atomic:** moved numeric facet input placeholders above their inputs ([#2155](https://github.com/coveo/ui-kit/issues/2155)) ([017fd23](https://github.com/coveo/ui-kit/commit/017fd23d4d155b5e88a98d17db5e58571b953a6d))
+
+
+### Features
+
+* **atomic:** additional control over outgoing i18n locales requests ([#2145](https://github.com/coveo/ui-kit/issues/2145)) ([b180cca](https://github.com/coveo/ui-kit/commit/b180cca801262d7589b41e32d814a9a4431f2da5))
+
+
+
+
+
+## [1.76.2](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.76.1...@coveo/atomic@1.76.2) (2022-06-24)
+
+
+### Bug Fixes
+
+* **atomic:** fix arrow navigation on input ([#2152](https://github.com/coveo/ui-kit/issues/2152)) ([54efce7](https://github.com/coveo/ui-kit/commit/54efce7ff2a6f672fcfb437908087ace3489b3d1))
+* **atomic:** improve click interaction in instant result ([#2129](https://github.com/coveo/ui-kit/issues/2129)) ([f7df573](https://github.com/coveo/ui-kit/commit/f7df57332f90941aee2be85224c6d1dd79c01a70))
+
+
+
+
+
+## [1.76.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.76.0...@coveo/atomic@1.76.1) (2022-06-23)
+
+
+### Bug Fixes
+
+* **atomic:** actually include fieldsToInclude in search from interface ([#2149](https://github.com/coveo/ui-kit/issues/2149)) ([1675f33](https://github.com/coveo/ui-kit/commit/1675f33bb35b8ba56f433a23ccb6784446ab1eec))
+* **atomic:** changed focus style of primary buttons ([#2143](https://github.com/coveo/ui-kit/issues/2143)) ([60bdf73](https://github.com/coveo/ui-kit/commit/60bdf73ceb2682102fb9e39f192b9fade838523e))
+* **atomic:** improved aria labels in facets ([#2144](https://github.com/coveo/ui-kit/issues/2144)) ([9f2a51b](https://github.com/coveo/ui-kit/commit/9f2a51baf725ccaf288e8b912af2d4aa480878ea))
+
+
+
+
+
+# [1.76.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.75.6...@coveo/atomic@1.76.0) (2022-06-22)
+
+
+### Bug Fixes
+
+* **atomic:** add region defaults ([#2139](https://github.com/coveo/ui-kit/issues/2139)) ([723e834](https://github.com/coveo/ui-kit/commit/723e834ed65732e09e8c26315fd9c9edacf1db0c))
+* **atomic:** hid modals when they were never opened ([#2132](https://github.com/coveo/ui-kit/issues/2132)) ([1801620](https://github.com/coveo/ui-kit/commit/18016201cdcf5d8fdefbff9ac0aec0e6c9e7e41e))
+* **atomic:** improved contrast between the stars and their background ([#2123](https://github.com/coveo/ui-kit/issues/2123)) ([4b8dfa0](https://github.com/coveo/ui-kit/commit/4b8dfa049934e229efeb7418f5e3919d8fe2db25))
+* **atomic:** make isCollapsed prop's value consistent ([#2138](https://github.com/coveo/ui-kit/issues/2138)) ([1dd0970](https://github.com/coveo/ui-kit/commit/1dd0970b3d43690ab5759e612397a2c6b655824f))
+* **atomic:** prevented additional refine modals being created when one already exists ([#2135](https://github.com/coveo/ui-kit/issues/2135)) ([c51fe4a](https://github.com/coveo/ui-kit/commit/c51fe4a29cb83d33780632dafa59c84fd31b5c79))
+
+
+### Features
+
+* **atomic:** add segmented scrollable component basic structure ([#2100](https://github.com/coveo/ui-kit/issues/2100)) ([7bb2d1c](https://github.com/coveo/ui-kit/commit/7bb2d1c17e60d4fd924f04b4b319fca3b62e3d8a))
+* **atomic:** allowed the page to be resized down to 320px ([#2136](https://github.com/coveo/ui-kit/issues/2136)) ([7461c43](https://github.com/coveo/ui-kit/commit/7461c43053ea24a5ee8ec2aec734c8fe435c4f69))
+* **atomic:** disableSearch prop for search box ([#2126](https://github.com/coveo/ui-kit/issues/2126)) ([a3d2c64](https://github.com/coveo/ui-kit/commit/a3d2c645ea904bf27254ce4eac4f63f4e8318de9))
+
+
+
+
+
 ## [1.75.6](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.75.5...@coveo/atomic@1.75.6) (2022-06-21)
 
 **Note:** Version bump only for package @coveo/atomic